custom hook useI18n can only be used in real components
