+++ packages/app-project/src/components/ProjectHeader/helpers/getNavLinks/getNavLinks.js
@@ -1,40 +1,30 @@
-// import counterpart from 'counterpart'
+import counterpart from 'counterpart'
 
-import { useI18n } from '@translations/i18n'
-
-function getNavLinks (isLoggedIn, baseUrl, defaultWorkflow, router) {
-  const { locale } = router
-  const { translate: t } = useI18n(locale, false)
-
+function getNavLinks (isLoggedIn, baseUrl, defaultWorkflow) {
   const classifyHref = defaultWorkflow ? `${baseUrl}/classify/workflow/${defaultWorkflow}` : `${baseUrl}/classify`
   const links = [
     {
       href: `${baseUrl}/about/research`,
-      text: t('about')
-      // text: counterpart('ProjectHeader.nav.about')
+      text: counterpart('ProjectHeader.nav.about')
     },
     {
       href: classifyHref,
-      text: t('classify')
-      // text: counterpart('ProjectHeader.nav.classify')
+      text: counterpart('ProjectHeader.nav.classify')
     },
     {
       href: `${baseUrl}/talk`,
-      text: t('talk')
-      // text: counterpart('ProjectHeader.nav.talk')
+      text: counterpart('ProjectHeader.nav.talk')
     },
     {
       href: `${baseUrl}/collections`,
-      text: t('collect')
-      // text: counterpart('ProjectHeader.nav.collect')
+      text: counterpart('ProjectHeader.nav.collect')
     }
   ]
 
   if (isLoggedIn) {
     links.push({
       href: `${baseUrl}/recents`,
-      text: t('recents')
-      // text: counterpart('ProjectHeader.nav.recents')
+      text: counterpart('ProjectHeader.nav.recents')
     })
   }
